LUCENE-4933: collapse Exact/SloppySimScorer into SimScorer
git-svn-id: https://svn.apache.org/repos/asf/lucene/dev/branches/branch_4x@1490974 13f79535-47bb-0310-9956-ffa450edef68

+++ solr/core/src/java/org/apache/solr/search/similarities/SweetSpotSimilarityFactory.java
@@ -180,7 +180,7 @@ public class SweetSpotSimilarityFactory extends DefaultSimilarityFactory {
   private static final class HyperbolicSweetSpotSimilarity 
     extends SweetSpotSimilarity {
     @Override
-    public float tf(int freq) {
+    public float tf(float freq) {
       return hyperbolicTf(freq);
     }
   };
